requirements.txt

firebase-admin
requests>=2.28.0
reportlab>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0
Flask-Login>=0.6.3
dnspython>=2.4.0
exifread>=3.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
python-whois>=0.9.0
google-generativeai
gunicorn
